Different entry point for Preact
