+++ karma.conf.js
@@ -60,6 +60,7 @@ function createEsbuildPlugin() {
 
 	const alias = {
 		deepsignal: subPkgPath("./packages/deepsignal"),
+		["deepsignal/preact"]: subPkgPath("./packages/deepsignal/preact"),
 	};
 
 	return {
@@ -252,7 +253,8 @@ module.exports = function (config) {
 
 		files: [
 			{
-				pattern: process.env.TESTS || "packages/*/test/**/*.test.tsx",
+				pattern:
+					process.env.TESTS || "packages/*/!(node_modules)/**/*.test.tsx",
 				watched: false,
 				type: "js",
 			},
@@ -263,7 +265,7 @@ module.exports = function (config) {
 		},
 
 		preprocessors: {
-			"packages/*/test/**/*": ["esbuild"],
+			"packages/*/!(node_modules)/**/*.test.tsx": ["esbuild"],
 		},
 
 		plugins: [
