+++ package.json
@@ -2,10 +2,12 @@
 	"name": "deepsignals",
 	"private": true,
 	"scripts": {
-		"prebuild": "rimraf packages/deepsignal/dist/",
-		"build": "pnpm build:deepsignal",
+		"prebuild": "rimraf packages/deepsignal/dist/ && rimraf packages/deepsignal/preact/dist",
+		"build": "pnpm build:deepsignal && pnpm build:deepsignal-preact",
 		"build:deepsignal": "microbundle --raw --cwd packages/deepsignal",
+		"build:deepsignal-preact": "microbundle --raw --cwd packages/deepsignal/preact && pnpm postbuild:deepsignal-preact",
 		"postbuild": "node ./scripts/node-13-exports.js",
+		"postbuild:deepsignal-preact": "cd packages/deepsignal/preact/dist && mv -f src/index.d.ts deepsignal-preact.d.ts && rm -dr preact && rm -dr src",
 		"lint": "eslint 'packages/**/*.{ts,tsx,js,jsx}'",
 		"test": "cross-env COVERAGE=true karma start karma.conf.js --single-run",
 		"test:minify": "cross-env COVERAGE=true MINIFY=true karma start karma.conf.js --single-run",
