+++ tsconfig.json
@@ -12,7 +12,8 @@
 		"stripInternal": true,
 		"noUnusedLocals": true,
 		"paths": {
-			"deepsignal": ["./packages/deepsignal/src/index.ts"]
+			"deepsignal": ["./packages/deepsignal/src/index.ts"],
+			"deepsignal/preact": ["./packages/deepsignal/preact/src/index.ts"]
 		}
 	}
 }
